FEAT convert netcdf tensor into int16 format save true_doy tensor

+++ src/openeo_mmdc/dataset/utils.py
@@ -147,7 +147,7 @@ def load_item_dataset_modality(
         ccp = ccp.compute()
         t_sel = ccp.where(ccp < max_pix_cc, drop=True)["t"]
         dataset = dataset.sel(t=t_sel)
-    return dataset
+    return dataset.astype("int16")
 
 
 def order_dataset_vars(dataset, list_vars_order=None):
